+++ app/smartbrain_3000/web/_app/immutable/entry/start.CJddLg6-.js
@@ -1,0 +1,1 @@
+import{a as e,n as t}from"../chunks/CFO5x2r6.js";export{e as load_css,t as start};