+++ app/smartbrain_3000/web/_app/immutable/entry/start.DwkzUSCd.js
@@ -1,1 +1,0 @@
-import{a as e,n as t}from"../chunks/CB3BI3v6.js";export{e as load_css,t as start};